Build the translations for all C++ programs and libraries.
buildlib/translations.cmake is introduced, featuring a macro which sets
up the targets for building the .mo files from the .po files in po/; split
in sub-domains such as libapt-pkg4.8.

There is no support for updating the po/apt-all.pot or po/*.po yet, this
will be in the next update.