* apt-pkg/acquire-item.cc:   - drop support for i18n/Index file (introduced in 0.8.11) and use     the Release file instead to get the Translations (Closes: #649314) * ftparchive/writer.cc:   - add 'Translation-*' to the default patterns
i18n/Index was never used outside debian - and even here it isn't used
consistently as only 'main' has such a file. As the Release file now
includes the Translation-* files we therefore drop support for i18n/Index.

A version supporting it was never part of a debian release and still
supporting it would mean that we get 99% of the time a 404 as response
to the request anyway and confuse archive maintainers who want to
provide all files APT tries to acquire.