fail in CopyFile if the FileFds have error flag set
Testing for global PendingErrors in users of CopyFile is incorrect
in so far as unrelated errors will prevent us from copying perfectly
fine files and checking for the validity of the files is just better
in CopyFiles as it already checks if files are at least opened.

Add also a higher-level error message to the error stack if it fails.