fix progress-segfault in case of dpkg errors/prompts
Errors and conffile prompts have a fourth information piece,
which the "old" code access which isn't provided by the "new" one.

This isn't checking if the messages are really well-formed,
so it could still segfault on misformed messages, but this code
needs more work anyway, so one step at a time.

Closes: 726047