reimplement rred to allow applying all the diffs in a single pass
Based on the idea presented in:
https://lists.debian.org/deity/2009/08/msg00169.html  and
https://lists.debian.org/debian-devel/2014/01/msg00081.html

It reads all patches one by one and merges them in-memory before
applying the merged changes to the index.

Beware: This commit by David Kalnischkies rips out the rred binary
rewrite unchanged (expect minor format issue corrections) from the
proposed changes, so this commit alone BREAKS pdiff completely.
The integration into the acquire system as it was prepared in the
previous POC will be done in the next commit to have proper 'blame'.