disable fnmatch()
The current PackageContainerInterface::FromString() will do a
FromFnmatch() first and then FromRegEx(). This commit reverts
that change to restore the old behavior to only look for RegEx
and not glob-style pattern. The rational is that:
a) currently a fnmatch() is misleadingly reported as a regex match to
   the user (Bug#738880)
b) a fnmatch may match something different than a a RegEx so the
   change broke a published interface