remove outdated README.MultiArch
Debian wheezy shipped MultiArch to the masses and the predictions
remained true in sofar as little changes in apt itself and many
other frontends were needed compared to the fallout if done differently.

The info included is this file is therefore no longer current and adds
no useful information anymore, so we can drop it for good.