(error) va_list 'args' was opened but not closed by va_end()
The manpage of va_start and co additionally says:
On some systems, va_end contains a closing '}' matching a '{' in
va_start, so that both macros must occur in the same function, and in a
way that allows this.

So instead of return/breaking instantly, we save the return, make a
proper turndown with va_end in all cases and only end after that.

Reported-By: cppcheck
Git-Dch: Ignore