document APT::Periodic::RandomSleep
The documentation in the patch is from
https://help.ubuntu.com/community/AutomaticSecurityUpdates

That page is licensed under Creative Commons Attribution-ShareAlike
3.0. Because I'm unsure how that license meshes with apt's license
I've not copied the text but formulated the same information freely
in my own words.

The original text was contributed by Chris Bainbridge [1][3]
and Kees Cook [2]. Thanks to them.

[1] https://help.ubuntu.com/community/AutomaticSecurityUpdates?action=diff&rev1=40&rev2=41
[2] https://help.ubuntu.com/community/AutomaticSecurityUpdates?action=diff&rev1=38&rev2=39
[3] https://help.ubuntu.com/community/AutomaticSecurityUpdates?action=diff&rev1=37&rev2=38

Closes: 776380
Thanks: Chris Bainbridge and Kees Cook for initial text