cleanup pdiff support detection decision
Its a bit unclean to create an item just to let the item decide that it
can't do anything and let it fail, so instead we let the item creator
decide in all cases if patching should be attempted.

Also pulls a small trick to get the hashes for the current file without
calculating them by looking at the 'old' Release file if we have it.

Git-Dch: Ignore