switch 'apt-mark hold' from Pkg to Ver based operation
Users hold a package foo (at version X) or try to prevent the
installation of foo (usually based on the information they know about
version X), even if we say that we "hold a package". Conceptionally we
also need to know about which architecture we are talking and that is an
information bound to a version (as a package can change architecture
over time).

We internally did this lookup from Pkg to Ver already, we just move this
to a central place where the user has a change to influence it now.

Git-Dch: Ignore